+++ app/test/e2e/ct-oauth/ct-oauth-facebook.spec.js
@@ -202,82 +202,81 @@ describe('Facebook auth endpoint tests', () => {
         confirmedUser.should.have.property('providerId').and.equal('10216001184997572');
     });
 
-    // it('Visiting /auth/facebook/callback while being logged in with an updated callbackUrl param should redirect to the new callback URL page', async () => {
-    //     if (skipTests) {
-    //         return;
-    //     }
-    //
-    //     const missingUser = await UserModel.findOne({ email: 'john.doe@vizzuality.com' }).exec();
-    //     should.not.exist(missingUser);
-    //
-    //     nock('https://graph.facebook.com')
-    //         .post('/oauth/access_token', {
-    //             grant_type: 'authorization_code',
-    //             redirect_uri: `${process.env.PUBLIC_URL}/auth/facebook/callback`,
-    //             client_id: process.env.TEST_FACEBOOK_OAUTH2_APP_ID,
-    //             client_secret: process.env.TEST_FACEBOOK_OAUTH2_APP_SECRET,
-    //             code: 'TEST_FACEBOOK_OAUTH2_CALLBACK_CODE'
-    //         })
-    //         .reply(200, {
-    //             access_token: 'facebook_access_token',
-    //             token_type: 'bearer',
-    //             expires_in: 5183974
-    //         });
-    //
-    //
-    //     nock('https://graph.facebook.com')
-    //         .get('/v2.5/me')
-    //         .query({
-    //             fields: 'id,name,picture,email',
-    //             access_token: 'facebook_access_token'
-    //         })
-    //         .reply(200, {
-    //             id: '10216001184997572',
-    //             name: 'John Doe',
-    //             picture: {
-    //                 data: {
-    //                     height: 50,
-    //                     is_silhouette: false,
-    //                     url: 'https://images.pexels.com/photos/20787/pexels-photo.jpg?auto=compress&cs=tinysrgb&h=750&w=1260',
-    //                     width: 50
-    //                 }
-    //             },
-    //             email: 'john.doe@vizzuality.com'
-    //         });
-    //
-    //     nock('https://www.wri.org')
-    //         .get('/')
-    //         .reply(200, 'ok');
-    //
-    //     // TODO: uncomment to reproduce current issue
-    //     // await requester
-    //     //     .get(`/auth?callbackUrl=https://www.google.com`);
-    //
-    //     await requester
-    //         .get(`/auth?callbackUrl=https://www.wri.org`);
-    //
-    //     const responseOne = await requester
-    //         .get(`/auth/facebook/callback?code=TEST_FACEBOOK_OAUTH2_CALLBACK_CODE`)
-    //         .redirects(0);
-    //
-    //     responseOne.should.redirect;
-    //     responseOne.should.redirectTo(new RegExp(`/auth/success$`));
-    //
-    //     const responseTwo = await requester
-    //         .get('/auth/success');
-    //
-    //     responseTwo.should.redirect;
-    //     responseTwo.should.redirectTo('https://www.wri.org/');
-    //
-    //     const confirmedUser = await UserModel.findOne({ email: 'john.doe@vizzuality.com' }).exec();
-    //     should.exist(confirmedUser);
-    //     confirmedUser.should.have.property('email').and.equal('john.doe@vizzuality.com');
-    //     confirmedUser.should.have.property('name').and.equal('John Doe');
-    //     confirmedUser.should.have.property('photo').and.equal('https://images.pexels.com/photos/20787/pexels-photo.jpg?auto=compress&cs=tinysrgb&h=750&w=1260');
-    //     confirmedUser.should.have.property('role').and.equal('USER');
-    //     confirmedUser.should.have.property('provider').and.equal('facebook');
-    //     confirmedUser.should.have.property('providerId').and.equal('10216001184997572');
-    // });
+    it('Visiting /auth/facebook/callback while being logged in with an updated callbackUrl param should redirect to the new callback URL page', async () => {
+        if (skipTests) {
+            return;
+        }
+
+        const missingUser = await UserModel.findOne({ email: 'john.doe@vizzuality.com' }).exec();
+        should.not.exist(missingUser);
+
+        nock('https://graph.facebook.com')
+            .post('/oauth/access_token', {
+                grant_type: 'authorization_code',
+                redirect_uri: `${process.env.PUBLIC_URL}/auth/facebook/callback`,
+                client_id: process.env.TEST_FACEBOOK_OAUTH2_APP_ID,
+                client_secret: process.env.TEST_FACEBOOK_OAUTH2_APP_SECRET,
+                code: 'TEST_FACEBOOK_OAUTH2_CALLBACK_CODE'
+            })
+            .reply(200, {
+                access_token: 'facebook_access_token',
+                token_type: 'bearer',
+                expires_in: 5183974
+            });
+
+
+        nock('https://graph.facebook.com')
+            .get('/v2.5/me')
+            .query({
+                fields: 'id,name,picture,email',
+                access_token: 'facebook_access_token'
+            })
+            .reply(200, {
+                id: '10216001184997572',
+                name: 'John Doe',
+                picture: {
+                    data: {
+                        height: 50,
+                        is_silhouette: false,
+                        url: 'https://images.pexels.com/photos/20787/pexels-photo.jpg?auto=compress&cs=tinysrgb&h=750&w=1260',
+                        width: 50
+                    }
+                },
+                email: 'john.doe@vizzuality.com'
+            });
+
+        nock('https://www.wri.org')
+            .get('/')
+            .reply(200, 'ok');
+
+        await requester
+            .get(`/auth?callbackUrl=https://www.google.com`);
+
+        await requester
+            .get(`/auth?callbackUrl=https://www.wri.org`);
+
+        const responseOne = await requester
+            .get(`/auth/facebook/callback?code=TEST_FACEBOOK_OAUTH2_CALLBACK_CODE`)
+            .redirects(0);
+
+        responseOne.should.redirect;
+        responseOne.should.redirectTo(new RegExp(`/auth/success$`));
+
+        const responseTwo = await requester
+            .get('/auth/success');
+
+        responseTwo.should.redirect;
+        responseTwo.should.redirectTo('https://www.wri.org/');
+
+        const confirmedUser = await UserModel.findOne({ email: 'john.doe@vizzuality.com' }).exec();
+        should.exist(confirmedUser);
+        confirmedUser.should.have.property('email').and.equal('john.doe@vizzuality.com');
+        confirmedUser.should.have.property('name').and.equal('John Doe');
+        confirmedUser.should.have.property('photo').and.equal('https://images.pexels.com/photos/20787/pexels-photo.jpg?auto=compress&cs=tinysrgb&h=750&w=1260');
+        confirmedUser.should.have.property('role').and.equal('USER');
+        confirmedUser.should.have.property('provider').and.equal('facebook');
+        confirmedUser.should.have.property('providerId').and.equal('10216001184997572');
+    });
 
     it('Visiting /auth/facebook/token with a valid Facebook OAuth token should generate a new token', async () => {
         await new UserModel({