+++ app/test/e2e/ct-oauth/ct-oauth-login-html.spec.js
@@ -179,77 +179,76 @@ describe('Auth endpoints tests', () => {
         responseTwo.should.redirectTo('https://www.wikipedia.org/');
     });
 
-    // it('Logging in successfully with /auth/login with callbackUrl should redirect to the callback page', async () => {
-    //     await createUserAndToken({
-    //         email: 'test@example.com',
-    //         role: 'ADMIN',
-    //         password: '$2b$10$1wDgP5YCStyvZndwDu2GwuC6Ie9wj7yRZ3BNaaI.p9JqV8CnetdPK',
-    //         salt: '$2b$10$1wDgP5YCStyvZndwDu2Gwu'
-    //     });
-    //
-    //     nock('https://www.wikipedia.org')
-    //         .get('/')
-    //         .reply(200, 'ok');
-    //
-    //     await requester
-    //         .get(`/auth/login?callbackUrl=https://www.wikipedia.org`);
-    //
-    //     const responseOne = await requester
-    //         .post(`/auth/login`)
-    //         .type('form')
-    //         .redirects(0)
-    //         .send({
-    //             email: 'test@example.com',
-    //             password: 'potato'
-    //         });
-    //
-    //     responseOne.should.redirect;
-    //     responseOne.should.redirectTo(new RegExp(`/auth/success$`));
-    //
-    //     const responseTwo = await requester
-    //         .get('/auth/success');
-    //
-    //     responseTwo.should.redirect;
-    //     responseTwo.should.redirectTo('https://www.wikipedia.org/');
-    // });
-    //
-    // it('Logging in successfully with /auth/login with an updated callbackUrl should redirect to the new callback page', async () => {
-    //     await createUserAndToken({
-    //         email: 'test@example.com',
-    //         role: 'ADMIN',
-    //         password: '$2b$10$1wDgP5YCStyvZndwDu2GwuC6Ie9wj7yRZ3BNaaI.p9JqV8CnetdPK',
-    //         salt: '$2b$10$1wDgP5YCStyvZndwDu2Gwu'
-    //     });
-    //
-    //     nock('https://www.wikipedia.org')
-    //         .get('/')
-    //         .reply(200, 'ok');
-    //
-    //     // TODO: uncomment to reproduce current issue
-    //     // await requester
-    //     //     .get(`/auth/login?callbackUrl=https://www.google.com`);
-    //
-    //     await requester
-    //         .get(`/auth/login?callbackUrl=https://www.wikipedia.org`);
-    //
-    //     const responseOne = await requester
-    //         .post(`/auth/login`)
-    //         .type('form')
-    //         .redirects(0)
-    //         .send({
-    //             email: 'test@example.com',
-    //             password: 'potato'
-    //         });
-    //
-    //     responseOne.should.redirect;
-    //     responseOne.should.redirectTo(new RegExp(`/auth/success$`));
-    //
-    //     const responseTwo = await requester
-    //         .get('/auth/success');
-    //
-    //     responseTwo.should.redirect;
-    //     responseTwo.should.redirectTo('https://www.wikipedia.org/');
-    // });
+    it('Logging in successfully with /auth/login with callbackUrl should redirect to the callback page', async () => {
+        await createUserAndToken({
+            email: 'test@example.com',
+            role: 'ADMIN',
+            password: '$2b$10$1wDgP5YCStyvZndwDu2GwuC6Ie9wj7yRZ3BNaaI.p9JqV8CnetdPK',
+            salt: '$2b$10$1wDgP5YCStyvZndwDu2Gwu'
+        });
+
+        nock('https://www.wikipedia.org')
+            .get('/')
+            .reply(200, 'ok');
+
+        await requester
+            .get(`/auth/login?callbackUrl=https://www.wikipedia.org`);
+
+        const responseOne = await requester
+            .post(`/auth/login`)
+            .type('form')
+            .redirects(0)
+            .send({
+                email: 'test@example.com',
+                password: 'potato'
+            });
+
+        responseOne.should.redirect;
+        responseOne.should.redirectTo(new RegExp(`/auth/success$`));
+
+        const responseTwo = await requester
+            .get('/auth/success');
+
+        responseTwo.should.redirect;
+        responseTwo.should.redirectTo('https://www.wikipedia.org/');
+    });
+
+    it('Logging in successfully with /auth/login with an updated callbackUrl should redirect to the new callback page', async () => {
+        await createUserAndToken({
+            email: 'test@example.com',
+            role: 'ADMIN',
+            password: '$2b$10$1wDgP5YCStyvZndwDu2GwuC6Ie9wj7yRZ3BNaaI.p9JqV8CnetdPK',
+            salt: '$2b$10$1wDgP5YCStyvZndwDu2Gwu'
+        });
+
+        nock('https://www.wikipedia.org')
+            .get('/')
+            .reply(200, 'ok');
+
+        await requester
+            .get(`/auth/login?callbackUrl=https://www.google.com`);
+
+        await requester
+            .get(`/auth/login?callbackUrl=https://www.wikipedia.org`);
+
+        const responseOne = await requester
+            .post(`/auth/login`)
+            .type('form')
+            .redirects(0)
+            .send({
+                email: 'test@example.com',
+                password: 'potato'
+            });
+
+        responseOne.should.redirect;
+        responseOne.should.redirectTo(new RegExp(`/auth/success$`));
+
+        const responseTwo = await requester
+            .get('/auth/success');
+
+        responseTwo.should.redirect;
+        responseTwo.should.redirectTo('https://www.wikipedia.org/');
+    });
 
     it('Logging in successfully with /auth/login with callbackUrl and token=true should redirect to the callback page and pass the token', async () => {
         await createUserAndToken({