+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # 12/12/2019
+- Change callback URL logic so that priority is:
+    - Query param
+    - Session
+    - Referer
 - Refactor tests to (even) better handle redirect cases
 
 # 11/12/2019